Restructure samples folder
