+++ javasteam-samples/src/main/java/in/dragonbra/javasteamsamples/_011_debuglog/SampleDebugLog.java
@@ -1,4 +1,4 @@
-package in.dragonbra.javasteamsamples._1logon;
+package in.dragonbra.javasteamsamples._011_debuglog;
 
 import in.dragonbra.javasteam.enums.EResult;
 import in.dragonbra.javasteam.steam.handlers.steamuser.LogOnDetails;
@@ -9,34 +9,37 @@ import in.dragonbra.javasteam.steam.steamclient.SteamClient;
 import in.dragonbra.javasteam.steam.steamclient.callbackmgr.CallbackManager;
 import in.dragonbra.javasteam.steam.steamclient.callbacks.ConnectedCallback;
 import in.dragonbra.javasteam.steam.steamclient.callbacks.DisconnectedCallback;
-import in.dragonbra.javasteam.util.log.DefaultLogListener;
+import in.dragonbra.javasteam.util.log.LogListener;
 import in.dragonbra.javasteam.util.log.LogManager;
 
 //
-// Sample 1: Logon
+// Sample 3: DebugLog
 //
-// the first act of business before being able to use steamkit2's features is to
-// log on to the steam network
+// sometimes is may be necessary to peek under the hood of SteamKit2
+// to debug or diagnose some issues
 //
-// interaction with steamkit is done through client message handlers and the results
-// come back through a callback queue controlled by a steamclient instance
+// to help with this, SK2 includes a component named the DebugLog
 //
-// your code must create a CallbackMgr, and instances of Callback<T>. Callback<T> maps a specific
-// callback type to a function, whilst CallbackMgr routes the callback objects to the functions that
-// you have specified. a Callback<T> is bound to a specific callback manager.
+// internal SK2 components will occasionally make use of the DebugLog
+// to share diagnostic information
 //
-// WARNING!
-// This the old login flow, we keep this sample around because it still currently works
-// for simple cases where you do not need to remember password.
-// See SampleLogonAuthentication for the new flow.
+// in order to use the DebugLog, a listener must first be registered with it
+//
+// by default, SK2 does not install any listeners, user code must install one
+//
+// additionally, the DebugLog is disabled by default in release builds,
+// but it may be enabled with the DebugLog.Enabled member
+//
+// you'll note that while this sample project is relatively similar to
+// Sample 1, the console output becomes very verbose
 //
 
 /**
  * @author lngtr
- * @since 2018-02-23
+ * @since 2021-10-11
  */
 @SuppressWarnings("FieldCanBeLocal")
-public class SampleLogon implements Runnable {
+public class SampleDebugLog implements Runnable {
 
     private SteamClient steamClient;
 
@@ -50,27 +53,51 @@ public class SampleLogon implements Runnable {
 
     private final String pass;
 
-    public SampleLogon(String user, String pass) {
+    // define our debuglog listener
+    static class MyListener implements LogListener {
+
+        // this function will be called when internal steamkit components write to the debuglog
+        @Override
+        public void onLog(Class<?> clazz, String message, Throwable throwable) {
+            // for this example, we'll print the output to the console
+            System.out.println("MyListener - " + clazz.getName() + ": " + message);
+        }
+
+        @Override
+        public void onError(Class<?> clazz, String message, Throwable throwable) {
+            // for this example, we'll print errors the output to the console
+            System.err.println("MyListener - " + clazz.getName() + ": " + message);
+        }
+    }
+
+    public SampleDebugLog(String user, String pass) {
         this.user = user;
         this.pass = pass;
     }
 
     public static void main(String[] args) {
         if (args.length < 2) {
-            System.out.println("Sample1: No username and password specified!");
+            System.out.println("Sample3: No username and password specified!");
             return;
         }
 
-        LogManager.addListener(new DefaultLogListener());
+        // JavaSteam already has one built in for convenience.
+        // LogManager.addListener(new DefaultLogListener());
 
-        new SampleLogon(args[0], args[1]).run();
+        LogManager.addListener(new MyListener());
+
+        new SampleDebugLog(args[0], args[1]).run();
     }
 
     @Override
     public void run() {
-
         // create our steamclient instance
         steamClient = new SteamClient();
+
+        // uncomment this if you'd like to dump raw sent and received packets
+        // that can be opened for analysis in NetHookAnalyzer
+        // NOTE: dumps may contain sensitive data (such as your Steam password)
+        // steamClient.setDebugNetworkListener(new NetHookNetworkListener());
 
         // create the callback manager which will route callbacks to function calls
         manager = new CallbackManager(steamClient);
@@ -124,7 +151,7 @@ public class SampleLogon implements Runnable {
     private void onLoggedOn(LoggedOnCallback callback) {
         if (callback.getResult() != EResult.OK) {
             if (callback.getResult() == EResult.AccountLogonDenied) {
-                // if we recieve AccountLogonDenied or one of its flavors (AccountLogonDeniedNoMailSent, etc.)
+                // if we receive AccountLogonDenied or one of its flavors (AccountLogonDeniedNoMailSent, etc.)
                 // then the account we're logging into is SteamGuard protected
                 // see sample 5 for how SteamGuard can be handled
                 System.out.println("Unable to logon to Steam: This account is SteamGuard protected.");
@@ -133,7 +160,7 @@ public class SampleLogon implements Runnable {
                 return;
             }
 
-            System.out.println("Unable to logon to Steam: " + callback.getResult() + " / " + callback.getExtendedResult());
+            System.out.println("Unable to logon to Steam: " + callback.getResult());
 
             isRunning = false;
             return;
